# knowledge/docs/api/api_spec.yaml
openapi: 3.0.3
info:
  title: 预测性智能交易系统 API
  description: |
    智能交易系统的 RESTful API 规范
    
    ## 认证方式
    所有 API 请求需要在 Header 中包含 JWT Token:
    ```
    Authorization: Bearer <token>
    ```
    
    ## 速率限制
    - 普通端点：100 请求/分钟
    - 交易端点：20 请求/分钟
    - 数据端点：200 请求/分钟
    
    ## 错误码
    - 400: 请求参数错误
    - 401: 未授权
    - 403: 禁止访问
    - 404: 资源不存在
    - 429: 请求超限
    - 500: 服务器内部错误
  version: 1.0.0
  contact:
    name: API Support
    email: support@trading-system.com
servers:
  - url: https://api.trading-system.com/v1
    description: 生产环境
  - url: https://staging-api.trading-system.com/v1
    description: 测试环境
  - url: http://localhost:8000/v1
    description: 本地开发
...
paths:
...
  /config/system:
    get:
      tags:
        - 配置
      summary: 获取系统配置
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfig'
    
    put:
      tags:
        - 配置
      summary: 更新系统配置
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemConfig'
      responses:
        '200':
          description: 更新成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RiskLimits:
      type: object
      properties:
        max_position_size:
          type: number
          default: 0.2
          description: 单笔最大仓位比例
        max_total_exposure:
          type: number
          default: 0.5
          description: 总仓位上限比例
        max_leverage:
          type: integer
          default: 10
          description: 最大杠杆倍数
        max_daily_loss:
          type: number
          default: 0.05
          description: 日最大亏损比例
        max_drawdown:
          type: number
          default: 0.15
          description: 最大回撤比例
        circuit_breaker_loss:
          type: number
          default: 0.03
          description: 熔断亏损阈值
        max_positions:
          type: integer
          default: 10
          description: 最大持仓数量
...
    SystemConfig:
      type: object
      properties:
        trading_enabled:
          type: boolean
        auto_execute_enabled:
          type: boolean
        min_signal_confidence:
          type: number
        default_leverage:
          type: number
        risk_limits:
          $ref: '#/components/schemas/RiskLimits'
        signal_weights:
          type: object
          properties:
            event:
              type: number
            ml:
              type: number
            sentiment:
              type: number
            technical:
              type: number
        notification_settings:
          type: object
          properties:
            email_enabled:
              type: boolean
            telegram_enabled:
              type: boolean
            alert_severity_threshold:
              type: string
        updated_at:
          type: string
          format: date-time
...